fix(api): give every capability a kebab-case id
Capability ids are public: the CLI derives its command path from them and
MCP's `invoke_capability` takes them verbatim. `deriveCapabilityId` suffixes
a NAMED export's identifier onto the path-derived id, so the seven capability
endpoints that were exported by name instead of as their file's default
shipped ids whose final segment was an internal handler function name, e.g.
`workspaces.anonymization-terms.deleteWorkspaceAnonymizationTerm`.

Give each of those endpoints its own kebab-case-named file with a default
export, so their ids derive purely from the handler path like the other 273.

BREAKING CHANGE: seven public capability ids are renamed.

  entities.read-summaries.readEntitySummariesCount
    -> entities.read-summaries-count
  workspaces.anonymization-allowlist.createWorkspaceAnonymizationAllowlistEntry
    -> workspaces.anonymization-allowlist.create
  workspaces.anonymization-allowlist.deleteWorkspaceAnonymizationAllowlistEntry
    -> workspaces.anonymization-allowlist.delete
  workspaces.anonymization-allowlist.readWorkspaceAnonymizationAllowlist
    -> workspaces.anonymization-allowlist.read
  workspaces.anonymization-terms.createWorkspaceAnonymizationTerms
    -> workspaces.anonymization-terms.create
  workspaces.anonymization-terms.deleteWorkspaceAnonymizationTerm
    -> workspaces.anonymization-terms.delete
  workspaces.anonymization-terms.readWorkspaceAnonymizationTerms
    -> workspaces.anonymization-terms.read

HTTP routes are unchanged; only the capability ids (and the CLI commands they
generate) move. The old ids had no CLI command of their own, so the rename
adds six commands rather than moving them.

Class guard, so this cannot recur: `isWellFormedCapabilityId` requires every
`.`-separated segment to be lowercase kebab-case. The catalog export fails
with an actionable error on any malformed id, the dispatch-module serializer
reuses the same pattern as its second layer, and a test asserts the shape over
the committed CLI snapshot.
